Implemented a ByteArray type with manual memory management that slab allocates and reuses chunks of bytes Changed all big memory processes like file-read, rollsum, compression and HashboxBlocks data handling to use ByteArray on both Client and Server Changed block compression so that not both uncompressed and compressed data needs to be in memory at the same time Total remake of the sendQueue in the client to pipeline compression and sending better Total remake of the incremental reference engine, it now uses a download worker process to concurrently load the last state from the server Changed a lot of mutex locks to atomic instructions Changed the roll-sum to not be used on files smaller than twice the minimum block size (128kb) Added -ignore option to client store, with default ignores for Mac and Unix platforms Added -index only option to storage GC Fixed a bug in storage GC that corrupted storage Fixed so that indexes are moved between idx files during GC if needed Fixed lockFile to work on both Windows and Unix as they use different methods for checking running processes

+++ core/protocol_test.go
@@ -178,7 +178,7 @@ func TestMessageSerialization(t *testing.T) {
 		}
 	}
 
-	if ok, dump := protocolPipeCompare(MsgTypeAccountInfo&MsgTypeServerMask, &MsgServerAccountInfo{[]Dataset{Dataset{Name: randomMsgString(), Size: uint64(rand.Int63()), ListH: randomByte128()}}}); !ok {
+	if ok, dump := protocolPipeCompare(MsgTypeAccountInfo&MsgTypeServerMask, &MsgServerAccountInfo{[]Dataset{Dataset{Name: randomMsgString(), Size: rand.Int63(), ListH: randomByte128()}}}); !ok {
 		t.Errorf("MsgServerAccountInfo did not pass serialization / deserialization test:\n%s", dump)
 	} else {
 		if testing.Verbose() {
@@ -199,8 +199,8 @@ func TestMessageSerialization(t *testing.T) {
 			DatasetState{
 				StateID:    randomByte128(),
 				BlockID:    randomByte128(),
-				Size:       uint64(rand.Int63()),
-				UniqueSize: uint64(rand.Int63()),
+				Size:       rand.Int63(),
+				UniqueSize: rand.Int63(),
 			},
 		},
 		ListH: randomByte128()}); !ok {
@@ -211,7 +211,7 @@ func TestMessageSerialization(t *testing.T) {
 		}
 	}
 
-	if ok, dump := protocolPipeCompare(MsgTypeAddDatasetState, &MsgClientAddDatasetState{AccountNameH: randomByte128(), DatasetName: randomMsgString(), State: DatasetState{StateID: randomByte128(), BlockID: randomByte128(), Size: uint64(rand.Int63()), UniqueSize: uint64(rand.Int63())}}); !ok {
+	if ok, dump := protocolPipeCompare(MsgTypeAddDatasetState, &MsgClientAddDatasetState{AccountNameH: randomByte128(), DatasetName: randomMsgString(), State: DatasetState{StateID: randomByte128(), BlockID: randomByte128(), Size: rand.Int63(), UniqueSize: rand.Int63()}}); !ok {
 		t.Errorf("MsgClientAddDatasetState did not pass serialization / deserialization test:\n%s", dump)
 	} else {
 		if testing.Verbose() {
